requirements.txt
appdirs==1.4.4
autopep8==1.5.5
backcall==0.2.0
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
colorama==0.4.4
configobj==5.0.6
cryptography==3.4.5
cx-Oracle==8.1.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kafka-python==2.0.2
kiwisolver==1.3.1
matplotlib==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
parso==0.8.1
pathspec==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==22.0.3
regex==2020.11.13
toml==0.10.2
tornado==6.1
tox==3.22.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.2
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
